refactor(issue-133): lift wireTetherFor into physics/Tether.ts
Helper was a sibling of PhysicsCard but operates on PhysicsWorld and
PhysicsTether — a physics-side concern, not a card-author concern. Moving
it now (before the upcoming card/ extraction) removes one cross-folder
edge that would otherwise outlive the refactor.

ParentKind alias renamed to TetherParentKind for namespace clarity.

Refs #133.
